Tile support to show travel exclusions. Added status graphic for poisoned monsters. Changing tile storage type from unsigned short to unsigned int to have room for more flag bits.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3300 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/externs.h
@@ -1360,12 +1360,12 @@ public:
 
 #ifdef USE_TILE
     // indexed by grid coords
-    FixedArray<unsigned short,GXM, GYM> tile_bk_fg; // tile fg
-    FixedArray<unsigned short,GXM, GYM> tile_bk_bg; // tile bg
+    FixedArray<unsigned int,GXM, GYM> tile_bk_fg; // tile fg
+    FixedArray<unsigned int,GXM, GYM> tile_bk_bg; // tile bg
     FixedArray<TileFlavor, GXM, GYM> tile_flavor;
     // indexed by (show-1) coords
-    FixedArray<unsigned short,ENV_SHOW_DIAMETER-2,ENV_SHOW_DIAMETER-2> tile_fg;
-    FixedArray<unsigned short,ENV_SHOW_DIAMETER-2,ENV_SHOW_DIAMETER-2> tile_bg;
+    FixedArray<unsigned int,ENV_SHOW_DIAMETER-2,ENV_SHOW_DIAMETER-2> tile_fg;
+    FixedArray<unsigned int,ENV_SHOW_DIAMETER-2,ENV_SHOW_DIAMETER-2> tile_bg;
 #endif
 
     FixedVector< cloud_struct, MAX_CLOUDS >  cloud; // cloud list
